move write_finished_file to writer
