+++ spinn_front_end_common/data/fec_data_writer.py
@@ -138,6 +138,17 @@ class FecDataWriter(PacmanDataWriter, SpiNNManDataWriter, FecDataView):
             except OSError:
                 time.sleep(0.5)
 
+    def write_finished_file(self) -> None:
+        """
+        Write a finished file to flag that the code has finished cleanly
+
+        This file signals the report directory can be removed.
+        """
+        finished_file_name = os.path.join(
+            self.get_timestamp_dir_path(), self.FINISHED_FILENAME)
+        with open(finished_file_name, "w", encoding="utf-8") as f:
+            f.writelines(self._get_timestamp())
+
     def set_allocation_controller(self, allocation_controller: Optional[
             MachineAllocationController]):
         """
